Refactoring to get rid of duplicated code
