+++ api/internal/plugins/execplugin/execplugin_test.go
@@ -4,7 +4,6 @@
 package execplugin_test
 
 import (
-	"fmt"
 	"strings"
 	"testing"
 
@@ -17,7 +16,6 @@ import (
 	"sigs.k8s.io/kustomize/api/resmap"
 	"sigs.k8s.io/kustomize/api/resource"
 	valtest_test "sigs.k8s.io/kustomize/api/testutils/valtest"
-	"sigs.k8s.io/kustomize/api/types"
 )
 
 func TestExecPluginConfig(t *testing.T) {
@@ -89,109 +87,5 @@ metadata:
 		p.Args()[4] != "s/$BAR/bar baz/g" ||
 		p.Args()[5] != "\\ \\ \\ " {
 		t.Fatalf("unexpected arg array: %#v", p.Args())
-	}
-}
-
-func makeConfigMap(rf *resource.Factory, name, behavior string, hashValue *string) *resource.Resource {
-	r := rf.FromMap(map[string]interface{}{
-		"apiVersion": "v1",
-		"kind":       "ConfigMap",
-		"metadata":   map[string]interface{}{"name": name},
-	})
-	annotations := map[string]string{}
-	if behavior != "" {
-		annotations[BehaviorAnnotation] = behavior
-	}
-	if hashValue != nil {
-		annotations[HashAnnotation] = *hashValue
-	}
-	if len(annotations) > 0 {
-		r.SetAnnotations(annotations)
-	}
-	return r
-}
-
-func makeConfigMapOptions(rf *resource.Factory, name, behavior string, disableHash bool) *resource.Resource {
-	return rf.FromMapAndOption(map[string]interface{}{
-		"apiVersion": "v1",
-		"kind":       "ConfigMap",
-		"metadata":   map[string]interface{}{"name": name},
-	}, &types.GeneratorArgs{
-		Behavior: behavior,
-		Options:  &types.GeneratorOptions{DisableNameSuffixHash: disableHash}})
-}
-
-func strptr(s string) *string {
-	return &s
-}
-
-func TestUpdateResourceOptions(t *testing.T) {
-	p := NewExecPlugin("")
-	if err := p.ErrIfNotExecutable(); err == nil {
-		t.Fatalf("expected unexecutable error")
-	}
-	rf := resource.NewFactory(kunstruct.NewKunstructuredFactoryImpl())
-	in := resmap.New()
-	expected := resmap.New()
-	cases := []struct {
-		behavior  string
-		needsHash bool
-		hashValue *string
-	}{
-		{hashValue: strptr("false")},
-		{hashValue: strptr("true"), needsHash: true},
-		{behavior: "replace"},
-		{behavior: "merge"},
-		{behavior: "create"},
-		{behavior: "nonsense"},
-		{behavior: "merge", hashValue: strptr("false")},
-		{behavior: "merge", hashValue: strptr("true"), needsHash: true},
-	}
-	for i, c := range cases {
-		name := fmt.Sprintf("test%d", i)
-		in.Append(makeConfigMap(rf, name, c.behavior, c.hashValue))
-		expected.Append(makeConfigMapOptions(rf, name, c.behavior, !c.needsHash))
-	}
-	actual, err := p.UpdateResourceOptions(in)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err.Error())
-	}
-	for i, a := range expected.Resources() {
-		b := actual.GetByIndex(i)
-		if b == nil {
-			t.Fatalf("resource %d missing from processed map", i)
-		}
-		if !a.Equals(b) {
-			t.Errorf("expected %v got %v", a, b)
-		}
-		if a.NeedHashSuffix() != b.NeedHashSuffix() {
-			t.Errorf("")
-		}
-		if a.Behavior() != b.Behavior() {
-			t.Errorf("expected %v got %v", a.Behavior(), b.Behavior())
-		}
-	}
-}
-
-func TestUpdateResourceOptionsWithInvalidHashAnnotationValues(t *testing.T) {
-	p := NewExecPlugin("")
-	if err := p.ErrIfNotExecutable(); err == nil {
-		t.Fatalf("expected unexecutable error")
-	}
-	rf := resource.NewFactory(kunstruct.NewKunstructuredFactoryImpl())
-	cases := []string{
-		"",
-		"FaLsE",
-		"TrUe",
-		"potato",
-	}
-	for i, c := range cases {
-		name := fmt.Sprintf("test%d", i)
-		in := resmap.New()
-		in.Append(makeConfigMap(rf, name, "", &c))
-		_, err := p.UpdateResourceOptions(in)
-		if err == nil {
-			t.Errorf("expected error from value %q", c)
-		}
 	}
 }
